muhasebe yönetimine başlandı

+++ KareAjans.Model/AccountingItemDTO.cs
@@ -8,7 +8,7 @@ namespace KareAjans.Model
     {
         public string ModelEmployeeName { get; set; }
         public string ProfessionalDegreeTitle { get; set; }
-        public byte NumberOfDays { get; set; }
+        public int NumberOfDays { get; set; }
         public bool IsLocal { get; set; }
         public decimal Wage { get; set; }
         public decimal Spent { get; set; }
